requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.1.0
async-timeout==4.0.3
attrs==23.1.0
beautifulsoup4==4.12.2
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
clipboard==0.0.4
colorama==0.4.6
dataclasses-json==0.6.3
Deprecated==1.2.14
diskcache==5.6.3
distro==1.8.0
exceptiongroup==1.2.0
faiss-cpu==1.7.4
filelock==3.13.1
frozenlist==1.4.0
fsspec==2023.12.1
gpt4all==2.0.2
greenlet==3.0.2
gTTS==2.4.0
h11==0.14.0
httpcore==1.0.2
httpx==0.25.2
huggingface-hub==0.19.4
idna==3.6
Jinja2==3.1.2
joblib==1.3.2
keyboard==0.13.5
llama-index==0.9.14.post3
llama_cpp_python==0.2.22
MarkupSafe==2.1.3
marshmallow==3.20.1
MouseInfo==0.1.3
mpmath==1.3.0
multidict==6.0.4
mypy-extensions==1.0.0
nest-asyncio==1.5.8
networkx==3.2.1
nltk==3.8.1
numpy==1.26.2
openai==1.3.9
packaging==23.2
pandas==2.1.4
Pillow==10.1.0
playsound==1.2.2
PyAudio==0.2.14
PyAutoGUI==0.9.54
pydantic==2.5.2
pydantic_core==2.14.5
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.30
python-dateutil==2.8.2
pytweening==1.0.7
pytz==2023.3.post1
PyYAML==6.0.1
regex==2023.10.3
requests==2.31.0
safetensors==0.4.1
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
SpeechRecognition==3.10.1
SQLAlchemy==2.0.23
sympy==1.12
tenacity==8.2.3
tiktoken==0.5.2
timm==0.9.8
torch==2.1.0
torchvision==0.16.0
tqdm==4.66.1
typing-inspect==0.9.0
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.1.0
wrapt==1.16.0
yarl==1.9.4
